misc: slight tweak to the makefile to allow customizing the uname for the module path
This simply makes the "VM_UNAME" variable conditionally assigned which
allows explicitly setting the kernel module directory to use when
building the modules. This has the effect of enabling the invoker to
choose the target kernel version to build against whilst still falling
back to the current kernel version when "VM_UNAME" is not given.